+++ R/beta_div.r
@@ -50,9 +50,6 @@ V_UNIFRAC <- 5L
 #'   'variance_adjusted_unifrac', 'wave_hedges', 'weighted_unifrac')`. Flexible
 #'   matching is supported (see below). Programmatic access via
 #'   `list_metrics('beta')`.
-#'   
-#' @param ...  Additional options to pass through to the called function. I.e.
-#'   `tree`, `pairs`, `alpha`, or `cpus`.
 #' 
 #' @return A numeric vector.
 #' 
@@ -129,8 +126,21 @@ V_UNIFRAC <- 5L
 #'     # Generalized UniFrac distances
 #'     beta_div(ex_counts, 'GUniFrac', tree = ex_tree)
 #'     
-beta_div <- function (counts, metric, ...) {
-  match_metric(metric, div = 'beta')$func(counts = counts, ...)
+beta_div <- function (
+    counts, 
+    metric, 
+    norm        = 'percent', 
+    power       = 1.5, 
+    pseudocount = NULL, 
+    alpha       = 0.5, 
+    tree        = NULL, 
+    pairs       = NULL, 
+    cpus        = n_cpus() ) {
+  
+  metric <- match_metric(metric, div = 'beta')
+  args   <- mget(metric$params, environment())
+  
+  do.call(metric$func, args)
 }
 
 
@@ -267,7 +277,7 @@ aitchison <- function (counts, pseudocount = NULL, pairs = NULL, cpus = n_cpus()
 #  -log(sum(sqrt(x * y)))
 #' @export
 #' @rdname bdiv_functions
-bhattacharyya <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+bhattacharyya <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_BHATTACHARYYA, counts, pairs, cpus, NULL)
@@ -279,7 +289,7 @@ bhattacharyya <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()
 # sum(abs(x-y)) / sum(x+y)
 #' @export
 #' @rdname bdiv_functions
-bray <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+bray <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_BRAY, counts, pairs, cpus, NULL)
@@ -290,7 +300,7 @@ bray <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(abs(x-y) / (x+y))
 #' @export
 #' @rdname bdiv_functions
-canberra <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+canberra <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_CANBERRA, counts, pairs, cpus, NULL)
@@ -301,7 +311,7 @@ canberra <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  max(abs(x-y))
 #' @export
 #' @rdname bdiv_functions
-chebyshev <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+chebyshev <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_CHEBYSHEV, counts, pairs, cpus, NULL)
@@ -326,7 +336,7 @@ chord <- function (counts, pairs = NULL, cpus = n_cpus()) {
 #  sqrt(sum((abs(x - y) / (x + y)) ^ 2))  
 #' @export
 #' @rdname bdiv_functions
-clark <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+clark <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_CLARK, counts, pairs, cpus, NULL)
@@ -337,7 +347,7 @@ clark <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  2 * sum((x - y)^2 / (x + y)^2)
 #' @export
 #' @rdname bdiv_functions
-divergence <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+divergence <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_DIVERGENCE, counts, pairs, cpus, NULL)
@@ -348,7 +358,7 @@ divergence <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sqrt(sum((x-y)^2))
 #' @export
 #' @rdname bdiv_functions
-euclidean <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+euclidean <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_EUCLIDEAN, counts, pairs, cpus, NULL)
@@ -361,7 +371,7 @@ euclidean <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(abs(x-y) / r) / n
 #' @export
 #' @rdname bdiv_functions
-gower <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+gower <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -375,7 +385,7 @@ gower <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sqrt(sum((sqrt(x) - sqrt(y)) ^ 2))
 #' @export
 #' @rdname bdiv_functions
-hellinger <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+hellinger <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -390,7 +400,7 @@ hellinger <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  1 - ((2 * sum(x * y)) / (z * sum(x) * sum(y)))
 #' @export
 #' @rdname bdiv_functions
-horn <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+horn <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_HORN, counts, pairs, cpus, NULL)
@@ -401,7 +411,7 @@ horn <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sqrt(sum(x * log(2 * x / (x+y)), y * log(2 * y / (x+y))) / 2)  
 #' @export
 #' @rdname bdiv_functions
-jensen <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+jensen <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -415,7 +425,7 @@ jensen <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(x * log(2 * x / (x+y)), y * log(2 * y / (x+y))) / 2  
 #' @export
 #' @rdname bdiv_functions
-jsd <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+jsd <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_JSD, counts, pairs, cpus, NULL)
@@ -426,7 +436,7 @@ jsd <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(log(1 + abs(x - y)))
 #' @export
 #' @rdname bdiv_functions
-lorentzian <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+lorentzian <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_LORENTZIAN, counts, pairs, cpus, NULL)
@@ -437,7 +447,7 @@ lorentzian <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(abs(x-y))
 #' @export
 #' @rdname bdiv_functions
-manhattan <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+manhattan <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_MANHATTAN, counts, pairs, cpus, NULL)
@@ -448,7 +458,7 @@ manhattan <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sqrt(sum((sqrt(x) - sqrt(y)) ^ 2))
 #' @export
 #' @rdname bdiv_functions
-matusita <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+matusita <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -463,7 +473,7 @@ matusita <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(abs(x - y)^p) ^ (1/p)
 #' @export
 #' @rdname bdiv_functions
-minkowski <- function (counts, rescale = TRUE, power = 1.5, pairs = NULL, cpus = n_cpus()) {
+minkowski <- function (counts, norm = 'percent', power = 1.5, pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_MINKOWSKI, counts, pairs, cpus, power)
@@ -489,7 +499,7 @@ morisita <- function (counts, pairs = NULL, cpus = n_cpus()) {
 #  sum(pmax(x, y)) / sum(x, y)
 #' @export
 #' @rdname bdiv_functions
-motyka <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+motyka <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_MOTYKA, counts, pairs, cpus, NULL)
@@ -500,7 +510,7 @@ motyka <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  2 * sum((x - y)^2 / (x + y))
 #' @export
 #' @rdname bdiv_functions
-psym_chisq <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+psym_chisq <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -514,7 +524,7 @@ psym_chisq <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(abs(x - y)) / sum(pmax(x, y))
 #' @export
 #' @rdname bdiv_functions
-soergel <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+soergel <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_SOERGEL, counts, pairs, cpus, NULL)
@@ -525,7 +535,7 @@ soergel <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum((x - y)^2 / (x + y))
 #' @export
 #' @rdname bdiv_functions
-squared_chisq <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+squared_chisq <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_SQUARED_CHISQ, counts, pairs, cpus, NULL)
@@ -536,7 +546,7 @@ squared_chisq <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()
 #  sum((sqrt(x) - sqrt(y)) ^ 2)
 #' @export
 #' @rdname bdiv_functions
-squared_chord <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+squared_chord <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_SQUARED_CHORD, counts, pairs, cpus, NULL)
@@ -547,7 +557,7 @@ squared_chord <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()
 #  sum((x-y)^2)
 #' @export
 #' @rdname bdiv_functions
-squared_euclidean <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+squared_euclidean <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -561,7 +571,7 @@ squared_euclidean <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cp
 #  sum(x * log(2 * x / (x+y)), y * log(2 * y / (x+y)))
 #' @export
 #' @rdname bdiv_functions
-topsoe <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+topsoe <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   
@@ -575,7 +585,7 @@ topsoe <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
 #  sum(abs(x - y) / pmax(x, y))
 #' @export
 #' @rdname bdiv_functions
-wave_hedges <- function (counts, rescale = TRUE, pairs = NULL, cpus = n_cpus()) {
+wave_hedges <- function (counts, norm = 'percent', pairs = NULL, cpus = n_cpus()) {
   
   validate_args()
   .Call(C_beta_div, BDIV_WAVE_HEDGES, counts, pairs, cpus, NULL)